requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
jira==0.40
MarkupSafe==0.23
oauthlib==0.7.2
ordereddict==1.1
python-dateutil==2.4.0
pytz==2014.10
redis==2.10.3
requests==2.5.2
requests-oauthlib==0.4.2
requests-toolbelt==0.3.1
six==1.9.0
tlslite==0.4.8
Werkzeug==0.10.1
